requirements.txt
--index-url https://download.pytorch.org/whl/cu121

torch
pandas
collection
python-box
pyyaml
tqdm
collection